fix to upgrade problem 0.23 -> 1.0.x assay_type_id
for some reason, assay_type_id is missing on assays for a 0.23 install. a migration exists that tries to remove it that fails, however worked on earlier installations.
Now added a check that the field exists as part of the migration.
same for technology_type_id

+++ db/migrate/20150625131437_remove_assay_and_tech_type_id_from_assay.rb
@@ -1,7 +1,11 @@
 class RemoveAssayAndTechTypeIdFromAssay < ActiveRecord::Migration
   def up
-    remove_column :assays,:assay_type_id
-    remove_column :assays, :technology_type_id
+    if column_exists? :assays, :assay_type_id
+      remove_column :assays,:assay_type_id
+    end
+    if column_exists? :assays, :technology_type_id
+      remove_column :assays, :technology_type_id
+    end
   end
 
   def down
